feat: add skeleton UI for maintenance and information page without styling
